OCaml 5.0 compatibility
From the latest failed build on `check.ocamllabs.io`:

```
OCaml's lib directory layout changed in 5.0. The unix subdirectory has been
automatically added to the search path, but you should add -I +unix to the
command-line to silence this alert (e.g. by adding unix to the list of
libraries in your dune file, or adding use_unix to your _tags file for
ocamlbuild, or using -package unix for ocamlfind).
```

It's not quite clear why this currently breaks the build, but this should
fix it anyway.